requirements.txt
fastapi
odmantic
pytest
pydantic_settings
python-dotenv
pytest-asyncio
PyJWT
passlib